rename pulse module
